pyproject.toml
[project]
name = "actionengine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bilibili-api-python>=16.3.0",
    "cohere>=5.13.5",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]
